Add files via upload
This Data Ingestion and Storage in AWS DynamoDB project is designed to showcase the process of ingesting data from a structured or semi-structured sources (csv and json), cleaning and transforming it on the fly with AWS Lambda python function and storing the cleaned data in S3 bucket as csv and json versions and as well storing it in an AWS DynamoDB database. 

DynamoDB is a fully managed, highly available, and scalable NoSQL database service provided by Amazon Web Services. 

The project involves the use of Python and the Boto3 library to interact with AWS services.